refactor: ViewModel에서 ShoppingApplication 의존성 제거 및 생성자 주입 방식으로 변경
