+++ app/src/main/java/woowacourse/shopping/feature/cart/CartViewModel.kt
@@ -13,6 +13,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import retrofit2.HttpException
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.repository.cart.CartRepository
@@ -23,20 +24,9 @@ import woowacourse.shopping.feature.common.state.ProductUiModel
 import java.io.IOException
 
 
-data class CartUiState(
-    val isLoading: Boolean = true,
-    val page: Int = 1,
-    val paginatedCartContents: List<CartItemUiModel> = emptyList(),
-    val checkMap: Map<Long, Boolean> = emptyMap(),
-    val totalPrice: Int = 0,
-    val totalCount: Int = 0,
-    val isFirstPage: Boolean = true,
-    val isLastPage: Boolean = true,
-)
-
 class CartViewModel(
     private val initialPageSize: Int = 5,
-    private val application: ShoppingApplication,
+    private val cartRepository: CartRepository
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(CartUiState())
     val uiState: StateFlow<CartUiState> = _uiState.asStateFlow()
@@ -44,12 +34,9 @@ class CartViewModel(
     private val _event = Channel<CartEvent>(Channel.BUFFERED)
     val event: Flow<CartEvent> = _event.receiveAsFlow()
 
-    lateinit var cartRepository: CartRepository
 
     init {
         viewModelScope.launch {
-            val appDependencies = application.appDependenciesDeferred.await()
-            cartRepository = appDependencies.cartRepository
             initialLoading()
         }
     }
@@ -333,7 +320,8 @@ class CartViewModel(
             viewModelFactory {
                 initializer {
                     val app = this[APPLICATION_KEY] as ShoppingApplication
-                    CartViewModel(5, app)
+                    val appDependencies = runBlocking { app.appDependenciesDeferred.await() }
+                    CartViewModel(5, appDependencies.cartRepository)
                 }
             }
     }
