+++ app/src/main/java/woowacourse/shopping/feature/productdetail/ProductDetailViewModel.kt
@@ -13,6 +13,7 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.repository.cart.CartRepository
 import woowacourse.shopping.data.repository.product.ProductRepository
@@ -20,24 +21,14 @@ import woowacourse.shopping.domain.Product
 import woowacourse.shopping.feature.common.state.ProductUiModel
 
 class ProductDetailViewModel(
-    private val application: ShoppingApplication,
+    private val productRepository: ProductRepository,
+    private val cartRepository: CartRepository,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(ProductDetailUiState())
     val uiState: StateFlow<ProductDetailUiState> = _uiState.asStateFlow()
 
     private val _event = MutableSharedFlow<AddEvent>()
     val event: SharedFlow<AddEvent> = _event.asSharedFlow()
-
-    lateinit var productRepository: ProductRepository
-    lateinit var cartRepository: CartRepository
-
-    init {
-        viewModelScope.launch {
-            val appDependencies = application.appDependenciesDeferred.await()
-            productRepository = appDependencies.productRepository
-            cartRepository = appDependencies.cartRepository
-        }
-    }
 
     fun initialLoading(
         productId: Long,
@@ -119,7 +110,11 @@ class ProductDetailViewModel(
             viewModelFactory {
                 initializer {
                     val app = this[APPLICATION_KEY] as ShoppingApplication
-                    ProductDetailViewModel(app)
+                    val appDependencies = runBlocking { app.appDependenciesDeferred.await() }
+                    ProductDetailViewModel(
+                        productRepository = appDependencies.productRepository,
+                        cartRepository = appDependencies.cartRepository,
+                    )
                 }
             }
     }
